refactor(inspections): the primitive keeps the file; its four callers move out
`inspection-core.service.ts` sat at 1131 against a 1132 cap, blocking
booking-deposit (#20)'s tier-2 fix. Pure move: no query, ordering, message,
log key or public signature changed, and 664 tests across
tests/unit/{inspections,usage}, the 85 tests in tests/unit/mcp (OpenAPI
snapshot byte-identical, not regenerated) and tests/workers/reinspections
pass untouched.

WHAT REMAINS IS THE PRIMITIVE — create one inspection, read one, and the
guards that creating one needs. Everything moved out is a CALLER of it, and
each one is composed as a field on the core service rather than promoted onto
the `InspectionService` facade. That is a deliberate constraint, not taste:
the facade is itself at 740/741, so four new fields and four constructor
lines would have broken its cap. Composing here leaves every existing
`this.core.x()` delegation on the facade byte-identical.

`inspection-query.service.ts` — reading MANY. The seam is cardinality:
`listInspections` and `getStats` answer questions about a SET and pay for it
with cursor pagination, a LEFT JOIN chain and a batched roster lookup.
Reading ONE loads the template and the results payload, which no list can
afford per row, so `getInspection` and `computePreflight` stay behind.

`inspection-reinspection.service.ts` — #119 rounds. The seam is the baseline:
both methods are meaningless without a PUBLISHED prior inspection, both read
its latest `report_versions` snapshot, and they must agree about what
`.original` means when the baseline is itself a re-inspection. That is why
they share a file and why `parseSnapshotData` — which had been sitting at
module scope in a file where only these two called it — goes with them.

`inspection-create-variants.service.ts` — the other ways an inspection comes
into existence. Neither `createFromWizard` nor `cloneInspection` is an
alternative to `createInspection`; each is a TRANSLATION into it, which is
why both call back into the core service instead of duplicating the insert,
and why the constructor takes it. `applyServicePriceOverrides` joins them as
the same shape of thing: a post-create hook the handler runs after an id
exists.

`inspection-recipients.service.ts` — who is attached, read two ways (the
Publish modal's deliverable list, the inspector portal's People card). They
share the source of truth and must not disagree about it: since Task 13
dropped the legacy contact columns, `inspection_people` is the only
persistence of who, and a second reader reaching for `inspections.client_name`
would resurrect GDPR-erased names.

BASELINE ENTRY LEFT ALONE AT 1132. The file reached 464, not under 400, so
its entry is neither removed (it would fail the 400 rule) nor tightened to
464 (that recreates the wall one line out). Getting the last 64 lines would
have meant carving the middle out of `createInspection` — a boundary chosen
to satisfy a counter, which is worse than the number. The practical result
stands: 668 lines of headroom where there was one.

`type-check:app` run explicitly, not just `:api`.
