+++ server/services/inspection/inspection-reinspection.service.ts
@@ -1,0 +1,288 @@
+import { eq, and, desc } from 'drizzle-orm';
+import { inspections, inspectionResults, users, tenantConfigs, reportVersions } from '../../lib/db/schema';
+import { PeopleService } from '../people.service';
+import { getRatingBucket, type RatingLevel } from '../../lib/report-utils';
+import { mapRatingSystemLevels } from '../../lib/map-rating-levels';
+import { logger } from '../../lib/logger';
+import { createPrimaryReport } from '../../lib/inspection/reports';
+import { findingKey, DEFAULT_UNIT } from '../../lib/finding-key';
+import { parseReinspectionStatuses, isOpenStatus } from '../../lib/reinspection-status';
+import { INSPECTION_STATUS } from '../../lib/status/inspection-status';
+import type { Inspection } from './shared';
+import type { ScopedDB } from '../../lib/db/scoped';
+import type { ImagesBinding } from '../../lib/media/strip-exif';
+import type { PlanQuotaGuard } from '../../features/plan-quota/guard';
+import { InspectionSubService } from './base';
+
+/** Parse a report_versions.snapshotJson payload (snapshotOnPublish serialises
+ *  `{ inspection, data, units }`); both re-inspection paths read only `.data`,
+ *  keyed by findingKey or legacy item id. */
+function parseSnapshotData(snapshotJson: string): { data?: Record<string, Record<string, unknown>> } {
+    return JSON.parse(snapshotJson) as { data?: Record<string, Record<string, unknown>> };
+}
+
+/**
+ * #119 — RE-INSPECTION ROUNDS: creating a follow-up round over a published
+ * baseline, and listing what is eligible to carry into one.
+ *
+ * The seam is the baseline. Both methods here are meaningless without a
+ * PUBLISHED prior inspection: `getReinspectCandidates` reads its latest
+ * `report_versions` snapshot to decide what is still open, and
+ * `createReinspection` seeds the new round's `inspection_results.data` from the
+ * same snapshot. Nothing else in the inspection lifecycle reads
+ * `report_versions` to write an inspection, and these two must agree about what
+ * `.original` means when a baseline is itself a re-inspection — which is why
+ * they share a file and a snapshot parser.
+ */
+export class InspectionReinspectionService extends InspectionSubService {
+    private readonly planQuota: PlanQuotaGuard | undefined;
+
+    constructor(
+        db: D1Database,
+        r2?: R2Bucket,
+        sdb?: ScopedDB,
+        kv?: KVNamespace,
+        images?: ImagesBinding,
+        planQuota?: PlanQuotaGuard,
+    ) {
+        super(db, r2, sdb, kv, images);
+        this.planQuota = planQuota;
+    }
+
+    /**
+     * #119 — Re-inspection. Creates a NEW draft inspection linked to a published
+     * baseline (the original OR a prior re-inspection). Seeds inspection_results.data
+     * for ONLY the selected items, each `{ original, followupStatus: null }`, where
+     * `original` carries the root finding forward from the baseline's latest published
+     * report_versions snapshot (or the propagated `.original` if the baseline is itself
+     * a re-inspection).
+     *
+     * GATE: the baseline must be published — i.e. have ≥1 report_versions row.
+     */
+    async createReinspection(
+        tenantId: string,
+        baselineId: string,
+        opts: { selectedItemIds: string[]; inspectorId?: string },
+    ): Promise<Inspection> {
+        const db = this.getDrizzle();
+
+        const baseline = await db.select().from(inspections)
+            .where(and(eq(inspections.id, baselineId), eq(inspections.tenantId, tenantId))).get();
+        if (!baseline) throw new Error('Baseline inspection not found');
+
+        const latestVersion = await db.select().from(reportVersions)
+            .where(and(eq(reportVersions.tenantId, tenantId), eq(reportVersions.inspectionId, baselineId)))
+            .orderBy(desc(reportVersions.versionNumber)).limit(1).get();
+        if (!latestVersion) throw new Error('Cannot re-inspect an unpublished baseline');
+
+        // When an explicit inspectorId is supplied, it MUST resolve to a user in
+        // this tenant. inspector_id has a DB FK to users.id; a foreign-tenant or
+        // bogus id would either violate the FK at runtime or assign the round to
+        // another tenant's user. Validate before use; omitted → baseline fallback.
+        if (opts.inspectorId) {
+            const owner = await db.select({ id: users.id }).from(users)
+                .where(and(eq(users.id, opts.inspectorId), eq(users.tenantId, tenantId))).get();
+            if (!owner) throw new Error('Inspector not found in this workspace');
+        }
+
+        const rootId = baseline.rootInspectionId ?? baseline.id;
+        const existingRounds = await db.select().from(inspections)
+            .where(and(eq(inspections.tenantId, tenantId), eq(inspections.rootInspectionId, rootId))).all();
+        const round = existingRounds.length + 1;
+
+        // The latest published snapshot is the carry-forward source. snapshotOnPublish
+        // serialises { inspection, data, units }; we read .data[itemId].
+        const baseSnapshot = parseSnapshotData(latestVersion.snapshotJson);
+        const baselineIsReinspection = baseline.sourceInspectionId != null;
+
+        const seeded: Record<string, unknown> = {};
+        for (const itemId of opts.selectedItemIds) {
+            const item = baseSnapshot.data?.[itemId] ?? {};
+            // When the baseline is itself a re-inspection AND its snapshot item already
+            // carries a propagated `.original` root finding, forward THAT (so round N
+            // always shows the root defect, never the intermediate follow-up state).
+            const original = baselineIsReinspection && item.original
+                ? item.original
+                : { rating: item.rating ?? null, notes: item.notes ?? null, photos: item.photos ?? [] };
+            seeded[itemId] = { original, followupStatus: null };
+        }
+
+        const id = crypto.randomUUID();
+        const createdAt = new Date();
+        // Quota is consumed only after every precondition check above (baseline
+        // existence, published-baseline gate, inspector ownership) has passed
+        // and immediately before the insert that actually creates the
+        // re-inspection — a failed validation must never burn a free tenant's
+        // lifetime slot.
+        await this.planQuota?.consumeInspection(tenantId);
+        await db.insert(inspections).values({
+            id,
+            tenantId,
+            // Reuse the baseline's property + client + template fields.
+            inspectorId:             opts.inspectorId ?? baseline.inspectorId ?? null,
+            propertyAddress:         baseline.propertyAddress,
+            addressPlaceId:          baseline.addressPlaceId,
+            addressStreet:           baseline.addressStreet,
+            addressCity:             baseline.addressCity,
+            addressState:            baseline.addressState,
+            addressZip:              baseline.addressZip,
+            addressCounty:           baseline.addressCounty,
+            addressLat:              baseline.addressLat,
+            addressLng:              baseline.addressLng,
+            templateId:              baseline.templateId,
+            templateSnapshot:        baseline.templateSnapshot,
+            templateSnapshotVersion: baseline.templateSnapshotVersion,
+            date:                    createdAt.toISOString(),
+            status:                  INSPECTION_STATUS.REQUESTED,
+            paymentStatus:           'unpaid',
+            price:                   0,
+            paymentRequired:         false,
+            agreementRequired:       false,
+            createdAt,
+            // #119 link columns.
+            sourceInspectionId: baselineId,
+            rootInspectionId:   rootId,
+            reinspectionRound:  round,
+        });
+
+        // Its own ORDER, so its own primary report — before the row naming it.
+        const primaryReportId = await createPrimaryReport(db, tenantId, id, null);
+
+        await db.insert(inspectionResults).values({
+            id:           crypto.randomUUID(),
+            tenantId,
+            inspectionId: id,
+            reportId:     primaryReportId,
+            data:         seeded as unknown as object,
+            lastSyncedAt: createdAt,
+        });
+
+        // Task 7c (people-role-profiles fix) — copy the baseline's
+        // inspection_people rows (client / buyer_agent / listing_agent / ...)
+        // onto the new re-inspection. Task 13 dropped the legacy
+        // clientContactId/clientName/clientEmail/clientPhone columns from the
+        // inspections row, so this copy is now the ONLY carry-forward of WHO.
+        // Without this, getInspection/listInspections (Task 9c-reads) resolve the client
+        // via inspection_people ONLY and would show a null client on every
+        // re-inspection. Non-fatal: a people-write failure must never roll
+        // back the already-committed re-inspection row.
+        try {
+            const people = new PeopleService({ DB: this.db });
+            const baselinePeople = await people.listPeople(tenantId, baselineId);
+            for (const p of baselinePeople) {
+                await people.addPerson(tenantId, id, p.contactId, p.roleProfileId);
+            }
+        } catch (err) {
+            logger.error('inspection-people copy from reinspection create failed', { inspectionId: id }, err instanceof Error ? err : undefined);
+        }
+
+        const created = await db.select().from(inspections).where(eq(inspections.id, id)).get();
+        return created as unknown as Inspection;
+    }
+
+    /**
+     * #119 (Task 6) — Candidate items for the "Create re-inspection" modal.
+     * Returns the baseline's still-open flagged items so the UI can pre-check
+     * the ones worth carrying forward. Computed off the SAME published snapshot
+     * `createReinspection` reads, so the returned `itemId`s are exactly the keys
+     * accepted as `selectedItemIds`.
+     *
+     * `open` default-check rule (mirrors the task spec):
+     *   - ORIGINAL baseline (no sourceInspectionId): item is open when its rating
+     *     bucket is `defect` or `monitor`.
+     *   - RE-INSPECTION baseline: item is open when its `followupStatus` is a
+     *     non-closed status (via isOpenStatus + the tenant's status set).
+     *
+     * Returns [] when the baseline is unpublished (no snapshot) — the caller
+     * gates the action on publication anyway, and the modal renders an empty
+     * state. Labels come from the baseline's templateSnapshot; an unmatched key
+     * degrades to the raw item id.
+     */
+    async getReinspectCandidates(
+        tenantId: string,
+        baselineId: string,
+    ): Promise<Array<{ itemId: string; label: string; originalNotes: string | null; open: boolean }>> {
+        const db = this.getDrizzle();
+
+        const baseline = await db.select().from(inspections)
+            .where(and(eq(inspections.id, baselineId), eq(inspections.tenantId, tenantId))).get();
+        if (!baseline) return [];
+
+        const latestVersion = await db.select().from(reportVersions)
+            .where(and(eq(reportVersions.tenantId, tenantId), eq(reportVersions.inspectionId, baselineId)))
+            .orderBy(desc(reportVersions.versionNumber)).limit(1).get();
+        if (!latestVersion) return [];  // unpublished baseline → no candidates
+
+        const baselineIsReinspection = baseline.sourceInspectionId != null;
+
+        // Snapshot data is keyed by findingKey (unit:section:item) or, for legacy
+        // inspections, the plain item id — the same keys createReinspection reads.
+        const snapData = parseSnapshotData(latestVersion.snapshotJson).data ?? {};
+
+        // Resolve item labels from the baseline's templateSnapshot (authoritative
+        // shape once an inspection exists). Both {sections:[...]} and flat-array
+        // formats are supported, matching getReportData's schema resolution.
+        const labelByItemId = new Map<string, string>();
+        const rawSnap = baseline.templateSnapshot as unknown;
+        const tplSnap = rawSnap
+            ? (typeof rawSnap === 'string' ? JSON.parse(rawSnap as string) : rawSnap)
+            : null;
+        const sections: Array<{ id?: string; items?: Array<Record<string, unknown>> }> = Array.isArray(tplSnap)
+            ? [{ id: 'general', items: tplSnap as Array<Record<string, unknown>> }]
+            : Array.isArray((tplSnap as { sections?: unknown })?.sections)
+                ? (tplSnap as { sections: Array<{ id?: string; items?: Array<Record<string, unknown>> }> }).sections
+                : [];
+        for (const sec of sections) {
+            for (const it of sec.items ?? []) {
+                const itemId = String(it.id ?? '');
+                if (!itemId) continue;
+                const label = String(it.label ?? it.title ?? it.name ?? itemId);
+                labelByItemId.set(itemId, label);
+                // Also map the composite findingKey so snapshot keys resolve.
+                labelByItemId.set(findingKey(DEFAULT_UNIT, String(sec.id ?? ''), itemId), label);
+            }
+        }
+
+        // Rating levels for bucket resolution (original-baseline rule). Read from
+        // the templateSnapshot.ratingSystem when present; absence degrades to the
+        // legacy string-bucket map inside getRatingBucket.
+        const snapLevels = !Array.isArray(tplSnap)
+            ? (tplSnap as { ratingSystem?: { levels?: unknown[] } } | null)?.ratingSystem?.levels
+            : undefined;
+        const levels: RatingLevel[] = Array.isArray(snapLevels)
+            ? mapRatingSystemLevels(snapLevels as Array<Record<string, unknown>>)
+            : [];
+
+        // Resolve the tenant's configured follow-up status set (re-inspection rule).
+        const configRow = await db.select({ reinspectionStatuses: tenantConfigs.reinspectionStatuses })
+            .from(tenantConfigs).where(eq(tenantConfigs.tenantId, tenantId)).get();
+        const resolvedStatuses = parseReinspectionStatuses(configRow?.reinspectionStatuses ?? null);
+
+        const out: Array<{ itemId: string; label: string; originalNotes: string | null; open: boolean }> = [];
+        for (const [itemId, entry] of Object.entries(snapData)) {
+            const rating = (entry.rating ?? null) as string | null;
+            const notes = (entry.notes ?? null) as string | null;
+            // A re-inspection snapshot may already carry the propagated root finding.
+            const original = (entry.original ?? null) as { notes?: string | null } | null;
+            const originalNotes = baselineIsReinspection && original ? (original.notes ?? null) : notes;
+
+            let open: boolean;
+            if (baselineIsReinspection) {
+                open = isOpenStatus((entry.followupStatus ?? null) as string | null, resolvedStatuses);
+            } else {
+                const bucket = getRatingBucket(rating, levels);
+                open = bucket === 'defect' || bucket === 'monitor';
+            }
+
+            out.push({
+                itemId,
+                label: labelByItemId.get(itemId) ?? itemId,
+                originalNotes,
+                open,
+            });
+        }
+        // Open items first, then by label — the pre-checked carry-forward set surfaces on top.
+        out.sort((a, b) => (a.open === b.open ? a.label.localeCompare(b.label) : a.open ? -1 : 1));
+        return out;
+    }
+}
